Add hardfork for proofs when filesize = 0
If the filesize is zero, storage proofs cannot be successful. You need
at least one byte to create a successful storage proof. This hardfork
allows for the filesize to be set to zero, allowing an empty (and
therefore much smaller) storage proof in situations where the host and
renter can terminate on friendly terms. This saves space on the
blockchain.